Re-factor scikit-learn converter
