+++ Makefile
@@ -8,7 +8,7 @@ endif
 lint:
 	wget -nc https://raw.githubusercontent.com/dmlc/dmlc-core/9db4b20c868341abe2a9fe52b652f7d9447ed406/scripts/lint.py
 	PYTHONPATH=./python:./runtime/python python lint.py treelite $(LINT_LANG) include src python tests/python \
-		--exclude_path python/treelite/gallery/sklearn --pylint-rc $(PWD)/python/.pylintrc
+		--pylint-rc $(PWD)/python/.pylintrc
 
 doxygen:
 	cd docs; doxygen
