+++ docs/tutorials/import.rst
@@ -64,13 +64,13 @@ Treelite.
 * :py:class:`sklearn.ensemble.GradientBoostingClassifier`
 
 To import scikit-learn models, use
-:py:meth:`treelite.gallery.sklearn.import_model`:
+:py:meth:`treelite.sklearn.import_model`:
 
 .. code-block:: python
 
   # clf is the model object generated by scikit-learn
-  import treelite.gallery.sklearn
-  model = treelite.gallery.sklearn.import_model(clf)
+  import treelite.sklearn
+  model = treelite.sklearn.import_model(clf)
 
 How about other packages?
 -------------------------
