+++ python/.pylintrc
@@ -1,6 +1,5 @@
 [MASTER]
 extension-pkg-whitelist=numpy
-ignore=treelite.gallery.sklearn
 
 disable=unexpected-special-method-signature,too-many-nested-blocks,useless-object-inheritance,import-outside-toplevel,unsubscriptable-object,attribute-defined-outside-init,unbalanced-tuple-unpacking
