+++ python/treelite/gallery/__init__.py
@@ -1,1 +1,7 @@
-# placeholder to get sphinx autodoc working
+# coding: utf-8
+"""Converter to ingest scikit-learn models into Treelite"""
+import warnings
+
+warnings.warn(('treelite.gallery.sklearn has been moved to treelite.sklearn. ' +
+               'treelite.gallery.sklearn will be removed in version 1.1.'),
+              FutureWarning)
