+++ tests/python/test_model_builder.py
@@ -5,7 +5,7 @@ import os
 import pytest
 import numpy as np
 import treelite
-import treelite.gallery.sklearn
+import treelite.sklearn
 import treelite_runtime
 from treelite.util import has_sklearn
 from treelite.contrib import _libext
@@ -127,7 +127,7 @@ def test_skl_converter_multiclass_classifier(tmpdir, clazz, toolchain):
     clf.fit(X, y)
     expected_prob = clf.predict_proba(X)
 
-    model = treelite.gallery.sklearn.import_model(clf)
+    model = treelite.sklearn.import_model(clf)
     assert model.num_feature == clf.n_features_
     assert model.num_output_group == clf.n_classes_
     assert (model.num_tree ==
@@ -168,7 +168,7 @@ def test_skl_converter_binary_classifier(tmpdir, clazz, toolchain):
     clf.fit(X, y)
     expected_prob = clf.predict_proba(X)[:, 1]
 
-    model = treelite.gallery.sklearn.import_model(clf)
+    model = treelite.sklearn.import_model(clf)
     assert model.num_feature == clf.n_features_
     assert model.num_output_group == 1
     assert model.num_tree == clf.n_estimators
@@ -207,7 +207,7 @@ def test_skl_converter_regressor(tmpdir, clazz, toolchain):  # pylint: disable=t
     clf.fit(X, y)
     expected_pred = clf.predict(X)
 
-    model = treelite.gallery.sklearn.import_model(clf)
+    model = treelite.sklearn.import_model(clf)
     assert model.num_feature == clf.n_features_
     assert model.num_output_group == 1
     assert model.num_tree == clf.n_estimators
